🚚 rename download-kubeconfig to just 'kubeconfig'
