+++ src/app/kubeconfig/kubeconfig-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { KubeconfigDownloadComponent } from './kubeconfig-download.component';
+import { KubeconfigDownloadComponent } from './kubeconfig.component';
 
 const routes: Routes = [{ path: '', component: KubeconfigDownloadComponent }];
