+++ src/app/kubeconfig/kubeconfig.component.html
