+++ src/app/kubeconfig/kubeconfig.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { KubeconfigDownloadComponent } from './kubeconfig-download.component';
+import { KubeconfigDownloadComponent } from './kubeconfig.component';
 
-import { KubeconfigDownloadRoutingModule } from './kubeconfig-download-routing.module';
+import { KubeconfigDownloadRoutingModule } from './kubeconfig-routing.module';
 import { SharedModule } from '../shared/shared.module';
 import { ReplaceValuesPipe } from './replace-values.pipe';
