+++ src/app/kubeconfig/replace-values.pipe.ts
